backend/pyproject.toml
[project]
name = "aido-backend"
version = "0.1.0"
description = "AIDO Todo Application Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.22",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.2.0",
    "email-validator>=2.2.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "black>=24.10.0",
    "flake8>=7.1.1",
    "mypy>=1.13.0",
    "pre-commit>=3.7.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
skip-string-normalization = true

[tool.flake8]
max-line-length = 88
extend-ignore = ['E203', 'W503']
exclude = ['.git', '__pycache__', '.venv', 'venv', 'dist', 'build']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"